SectionController cleanup
-SectionController code cleanup.
-Section text doesn't accept angle brackets.
-Created a StringKeys class to find connections between Controllers and Actions easier and store common strings.
-SelectSections actions use the new SectionModel View Model.
-Added some error handling to the SQL Server interface.

+++ TheWritersNet/TheWritersNet/Controllers/SectionController.cs
@@ -13,42 +13,46 @@ namespace TheWritersNet.Controllers
 {
     public class SectionController : Controller
     {
-        // GET: Section
-        public ActionResult Index()
-        {
-            return View();
-        }
-
         [Authorize]
         public ActionResult SelectSections(int pageID)
         {
             IDBConnector db = DBConnectorFactory.GetDBConnector();
-            List<DBSectionModel> sections = db.SelectPagePositions(pageID);
-            sections[0].PageID = pageID;
+            List<DBSectionModel> dbSections = db.SelectPagePositions(pageID);
 
-            return View(sections);
+            List<SectionModel> sections = new List<SectionModel>();
+            foreach (DBSectionModel section in dbSections)
+            {
+                section.PageID = pageID;
+                sections.Add(ConvertSectionModel(section));
+            }
+
+            return View(dbSections);
         }
 
         [Authorize]
         [HttpPost]
-        public ActionResult SelectSections(List<DBSectionModel> sections)
+        public ActionResult SelectSections(List<SectionModel> sections)
         {
-            List<DBSectionModel> removePositions = new List<DBSectionModel>();
-            List<DBSectionModel> addPositions = new List<DBSectionModel>();
-            foreach (DBSectionModel section in sections)
+            if (ModelState.IsValid)
             {
-                section.PageID = sections[0].PageID;
-                if (section.IsSelected)
-                    addPositions.Add(section);
-                else
-                    removePositions.Add(section);
+                List<DBSectionModel> removePositions = new List<DBSectionModel>();
+                List<DBSectionModel> addPositions = new List<DBSectionModel>();
+                foreach (SectionModel section in sections)
+                {
+                    if (section.IsSelected)
+                        addPositions.Add(ConvertSectionModel(section));
+                    else
+                        removePositions.Add(ConvertSectionModel(section));
+                }
+
+                IDBConnector db = DBConnectorFactory.GetDBConnector();
+                db.InsertPositions(addPositions);
+                db.DeletePositions(removePositions);
+
+                return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID = sections[0].PageID });
             }
 
-            IDBConnector db = DBConnectorFactory.GetDBConnector();
-            db.InsertPositions(addPositions);
-            db.DeletePositions(removePositions);
-
-            return RedirectToAction("EditFromID", "Page", new { pageID = sections[0].PageID });
+            return View(sections);
         }
 
         [Authorize]
@@ -56,10 +60,15 @@ namespace TheWritersNet.Controllers
         {
             IDBConnector db = DBConnectorFactory.GetDBConnector();
             DBSectionModel section = db.SelectSection(sectionID);
-            section.PageID = pageID;
-            section.Text = MarkdownConverter.MarkdownToHTML(section.Text, websiteID);
+            if (section != null)
+            {
+                section.PageID = pageID;
+                section.Text = MarkdownConverter.MarkdownToHTML(section.Text, websiteID);
 
-            return View(ConvertSectionModel(section));
+                return View(ConvertSectionModel(section));
+            }
+
+            return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID });
         }
 
         [Authorize]
@@ -81,7 +90,7 @@ namespace TheWritersNet.Controllers
                 List<SectionLinkModel> links = MarkdownConverter.FindInternalLinks(section.Text, index);
                 db.InsertSectionLinks(links);
 
-                return RedirectToAction("EditFromID", "Page", new { pageID = section.PageID });
+                return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID = section.PageID });
             }
 
             return View(section);
@@ -92,9 +101,14 @@ namespace TheWritersNet.Controllers
         {
             IDBConnector db = DBConnectorFactory.GetDBConnector();
             DBSectionModel section = db.SelectSection(sectionID);
-            section.PageID = pageID;
+            if (section != null)
+            {
+                section.PageID = pageID;
 
-            return View(ConvertSectionModel(section));
+                return View(ConvertSectionModel(section));
+            }
+
+            return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID });
         }
 
         [Authorize]
@@ -112,7 +126,7 @@ namespace TheWritersNet.Controllers
                 List<SectionLinkModel> links = MarkdownConverter.FindInternalLinks(section.Text, section.SectionID);
                 db.MergeSectionLinks(links, section.SectionID);
 
-                return RedirectToAction("EditFromID", "Page", new { pageID = section.PageID });
+                return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID = section.PageID });
             }
 
             return View(section);
@@ -123,9 +137,14 @@ namespace TheWritersNet.Controllers
         {
             IDBConnector db = DBConnectorFactory.GetDBConnector();
             DBSectionModel section = db.SelectSection(sectionID);
-            section.PageID = pageID;
+            if (section != null)
+            {
+                section.PageID = pageID;
 
-            return View(ConvertSectionModel(section));
+                return View(ConvertSectionModel(section));
+            }
+
+            return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID });
         }
 
         [Authorize]
@@ -135,7 +154,7 @@ namespace TheWritersNet.Controllers
             IDBConnector db = DBConnectorFactory.GetDBConnector();
             db.DeleteSection(section.SectionID);
 
-            return RedirectToAction("EditFromID", "Page", new { pageID = section.PageID });
+            return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID = section.PageID });
         }
 
         [Authorize]
@@ -145,19 +164,25 @@ namespace TheWritersNet.Controllers
             IDBConnector db = DBConnectorFactory.GetDBConnector();
             db.DeleteSectionFromPage(sectionID, pageID);
             
-            return RedirectToAction("EditFromID", "Page", new { pageID });
+            return RedirectToAction(StringKeys.EDIT_FROM_ID, StringKeys.PAGE_CONTROLLER, new { pageID });
         }
 
         private SectionModel ConvertSectionModel(DBSectionModel section)
         {
-            return new SectionModel()
+            SectionModel convertedSection = new SectionModel()
             {
                 PageID = section.PageID,
+                SectionID = section.SectionID,
                 DisplayTitle = section.DisplayTitle,
-                Position = section.Position,
                 Text = section.Text,
                 Title = section.Title
             };
+
+            int position;
+            int.TryParse(section.Position, out position);
+            convertedSection.Position = position;
+
+            return convertedSection;
         }
 
         private DBSectionModel ConvertSectionModel(SectionModel section)
@@ -167,7 +192,7 @@ namespace TheWritersNet.Controllers
                 PageID = section.PageID,
                 SectionID = section.SectionID,
                 DisplayTitle = section.DisplayTitle,
-                Position = section.Position,
+                Position = section.Position.ToString(),
                 Text = section.Text,
                 Title = section.Title
             };
